TD4 : Server récup dealabs P1

+++ server/websites/avenuedelabrique.js
@@ -11,7 +11,7 @@ const parse = data => {
 
   return $('div.prods a')
     .map((i, element) => {
-      const price = parseFloat(
+      const price = parseInt(
         $(element)
           .find('span.prodl-prix span')
           .text()
